Add files via upload
These two peices of code were aimed at running multiple HC-SR04 sensors simulatniously in order to record information on an unknown surface in front of it.
The first code used only two sensors and would record the distance and angle of the surface in 2 dimensions. With more time the code would have been programed to work for 3D as well and would have added a third sensor.
The second code used vector geometry to establish the overall direction of the surface of interest and then calculated its general distance from the sensors.